+++ test/request.test.ts
@@ -1,0 +1,5 @@
+describe('TDI.Request', () => {
+  it('it should send request by clicking on tdi button with a link', async () => {
+    // Mock fetch or use a test server
+  });
+});